+++ libbirch/libbirch.hpp
@@ -91,7 +91,7 @@ struct DefaultView<0> {
  * @param start First index.
  * @param end Last index.
  */
-Range<> make_range(const ptrdiff_t start, const ptrdiff_t end);
+Range<> make_range(const int64_t start, const int64_t end);
 
 /**
  * Make a frame, no arguments.
@@ -131,7 +131,7 @@ EmptyView make_view();
  *
  * @ingroup libbirch
  */
-template<ptrdiff_t offset_value, int64_t length_value>
+template<int64_t offset_value, int64_t length_value>
 auto make_view(const Range<offset_value,length_value>& arg) {
   auto head = arg;
   auto tail = make_view();
@@ -143,14 +143,14 @@ auto make_view(const Range<offset_value,length_value>& arg) {
  *
  * @ingroup libbirch
  */
-NonemptyView<Index<>,EmptyView> make_view(const ptrdiff_t arg);
+NonemptyView<Index<>,EmptyView> make_view(const int64_t arg);
 
 /**
  * Make a view, multiple arguments.
  *
  * @ingroup libbirch
  */
-template<ptrdiff_t offset_value, int64_t length_value, class ... Args>
+template<int64_t offset_value, int64_t length_value, class ... Args>
 auto make_view(const Range<offset_value,length_value>& arg, Args ... args) {
   auto head = arg;
   auto tail = make_view(args...);
@@ -163,7 +163,7 @@ auto make_view(const Range<offset_value,length_value>& arg, Args ... args) {
  * @ingroup libbirch
  */
 template<class ... Args>
-auto make_view(const ptrdiff_t arg, Args ... args) {
+auto make_view(const int64_t arg, Args ... args) {
   auto head = Index<mutable_value>(arg);
   auto tail = make_view(args...);
   return NonemptyView<decltype(head),decltype(tail)>(head, tail);
